+++ crates/widget/src/overlay.rs
@@ -1,2 +1,3 @@
 //! Display interactive elements on top of other widgets.
 pub mod menu;
+pub mod modal;